requirements/base.txt
psycopg2==2.6.1
django==1.9.4
djangorestframework==3.3.3
django-rest-auth==0.7.0
django-allauth==0.25.2
